Added help pages
I used an old website I had and transformed it into a Project-Rainbow help page! Feel free to edit it as you would like. The website links to downloads, forums, gives basic info, links to popular plugins, etc.

I hope this is of help! I spent a while working on it.